First upload of my work on the google homepage
This upload contains both HTML and CSS files which consist of the code for the entire project. however there are about 2 icons i haven't added due to the fact that i couldn't get them for free but i am working on them. Would work on them an them upload later and also add a commit message .